+++ packages/wave/CHANGELOG.md
@@ -1,10 +1,17 @@
 # @vastly/wave
 
+## 0.3.1
+
+### Patch Changes
+
+- Output credentials for appsync api
+
 ## 0.3.0
 
 ### Minor Changes
 
 - Update prisma-appsync microservice template
+
 ## 0.2.4
 
 ### Patch Changes